autoskill delete btn is on toolbar
