+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,8 @@
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
             android:name=".ui.AutoSkillItemActivity"
-            android:label="Edit AutoSkill Item" />
+            android:label="Edit AutoSkill Item"
+            android:theme="@style/AppTheme.NoActionBar"/>
         <activity
             android:name=".ui.ManageAutoSkillActivity"
             android:label="AutoSkill List" />
