+++ app/src/main/res/drawable/ic_delete.xml
@@ -3,5 +3,5 @@
         android:width="24dp"
         android:viewportWidth="24"
         android:viewportHeight="24">
-  <path android:fillColor="#e53935" android:pathData="M19,4H15.5L14.5,3H9.5L8.5,4H5V6H19M6,19A2,2 0 0,0 8,21H16A2,2 0 0,0 18,19V7H6V19Z" />
+  <path android:fillColor="#ef5350" android:pathData="M19,4H15.5L14.5,3H9.5L8.5,4H5V6H19M6,19A2,2 0 0,0 8,21H16A2,2 0 0,0 18,19V7H6V19Z" />
 </vector>
