+++ app/src/main/res/values/styles.xml
@@ -7,7 +7,10 @@
         <item name="android:textAllCaps">false</item>
     </style>
 
-    <style name="img_button_borderless" parent="Widget.AppCompat.Button.Borderless">
+    <style name="img_btn_delete" parent="Widget.AppCompat.Button.Borderless">
         <item name="android:minWidth">0dp</item>
+        <item name="android:layout_width">wrap_content</item>
+        <item name="android:layout_height">wrap_content</item>
+        <item name="android:src">@drawable/ic_delete</item>
     </style>
 </resources>
